Use a notifier instead of a never ending loop in GUI.
Removed the never ending loop in Client and replaced the isDone() method in
done with a doWait() (bad name) method instead. This causes the Client main
execution thread to pause until the play button is pressed. When the button
is pressed, a method called doNotify() wakes the Clients main execution
thread and the execution continues. This is still a hack, but a better one.