The server now waits for client init data.
The client init data containing car color etc is now added before the client
is added to the clienthandler. The Car is now added to the server before the
ClientConnection. This makes it impossible to request the clients car (which
doesn't exist) until the server is done adding it. The Car map in the server
needs to be synchronized if players are added during gameplay.