Checkpoints are now working. Both logic and draw.
Checkpoints are stored as Line2D array in the RaceCourse retreived from
raceCourse server. Checkpoints shall be intersected one by one, starting at
index 0. Intersection is controlled in Car.update. Car keeps track of
nextCheckpoint index. Index starts over when all checkpoints have been cleared.

To send the nextCheckpoint to clients, it was added to the CarUpdate,
meaning all nextCheckpoint data is sent to each client. To let the client
distinguish which data is relevant to itself, a clientID variable was added
to RaceUpdate, which is set by the ClientHandler before sending the message.
Since the clientID would be of no use with the old array of CarUpdate that
was sent, I switched its datatype to HashMap<Integer, CarUpdate> with
corresponfding IDs.

The client receives the nextCheckpoint index and uses it to draw the current
checkpoint goal with a different color in GameFrame.GamePanel.paint()